refactor: Move soap stuff to soap24 package
