+++ soap24/auth24/auth24.gen.go
