+++ soap24/client.go
@@ -1,22 +1,22 @@
 //go:generate go run generate24.go
 
-package so24
+package soap24
 
 import (
 	"fmt"
 
 	"github.com/hooklift/gowsdl/soap"
-	"github.com/valuetechdev/24sevenoffice-go/so24/account24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/attachment24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/auth24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/client24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/company24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/invoice24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/payroll24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/person24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/product24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/project24"
-	"github.com/valuetechdev/24sevenoffice-go/so24/transaction24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/account24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/attachment24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/auth24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/client24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/company24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/invoice24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/payroll24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/person24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/product24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/project24"
+	"github.com/valuetechdev/24sevenoffice-go/soap24/transaction24"
 )
 
 const (
