+++ soap24/generate24.go
